Refactor some field
