+++ src/components/HeaderWithImage/HeaderWithImage.tsx
@@ -1,6 +1,6 @@
 import React, { FC, FunctionComponent } from 'react';
 
-import Text from '../../common/components/fields/Text/Text'
+import Text from '../../common/components/fields/././text/Text'
 import Title from '../../components/Title/Title';
 
 import styles from './HeaderWithImage.module.css';
