+++ src/pages/home/Home.tsx
@@ -1,5 +1,5 @@
 /* test coverage not required */
-import Text from 'common/components/fields/Text/Text';
+import Text from 'common/components/fields/text/Text';
 import HeaderWithImage from 'components/HeaderWithImage/HeaderWithImage';
 import Title from 'components/Title/Title';
 import { AuthContext } from 'context/AuthContextProvider';
