+++ tests/common/components/dialogs/Modal/ModalBody.test.tsx
@@ -3,7 +3,7 @@ import '@testing-library/jest-dom';
 import React from 'react';
 
 import ModalBody from '../../../../../src/common/components/dialogs/Modal/ModalBody';
-import Text from '../../../../../src/common/components/fields/Text/Text';
+import Text from '../../../../../src/common/components/fields/././text/Text';
 
 describe('ModalBody', () => {
   it('it renders its children', () => {
